BasicIOEventLoop: create mtGateway non-lazily
Currently the BasicIOEventLoop creates EventLoopMTGateway lazily - when
its MT executor is used for the first time. The problem with this
approach is that the EventLoopMTGateway's constructor accesses the event
loop (it registers a watch for a pipe), so calling it from other threads
is an error. And first use of the MT executor is quite possibly from a
thread other that the one the event loop is running on.

Fix this by creating the EventLoopMTGateway instance directly in the
BasicIOEventLoop's constructor.